make sure client is defined by moving it to its own file, fetch description in the correct fashion, correctly send discordChannel in rosterService.updateScrimSignupCount
