+++ src/index.ts
@@ -1,17 +1,8 @@
 import * as path from "node:path";
 import * as fs from "node:fs";
-import { GatewayIntentBits } from "discord.js";
-import ExtendedClient from "./ExtendedClient";
+import { client } from "./ExtendedClient";
 import { commands } from "./commands";
 import { appConfig } from "./config";
-
-export const client = new ExtendedClient({
-  intents: [
-    GatewayIntentBits.Guilds,
-    GatewayIntentBits.GuildPresences,
-    GatewayIntentBits.GuildMembers,
-  ],
-});
 
 for (const command of commands) {
   client.commands.set(command.name, command);
